# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Multi-Institution Admission and Student Onboarding Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications
      parameters:
        - name: institutionId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Start a new admission application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/approve:
    post:
      operationId: approveApplication
      tags: [applications]
      summary: Approve application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/reject:
    post:
      operationId: rejectApplication
      tags: [applications]
      summary: Reject application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectionInput"
      responses:
        "200":
          description: Rejected application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/checkin:
    post:
      operationId: checkInStudent
      tags: [applications]
      summary: Check in student on reporting day
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checked in application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/verify-code:
    post:
      operationId: verifyAdmissionCode
      tags: [applications]
      summary: Verify admission letter code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionCodeVerification"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionVerificationResult"

  # ─── STUDENTS ─────────────────────────────────────────────────────────────────
...
  /allocations:
    get:
      operationId: listAllocations
      tags: [allocations]
      summary: List allocations for institution
      parameters:
        - name: institutionId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentAllocation"
    post:
      operationId: createAllocation
      tags: [allocations]
      summary: Assign student to class, stream, dormitory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocationInput"
      responses:
        "201":
          description: Allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentAllocation"
  /allocations/{id}:
    patch:
      operationId: updateAllocation
      tags: [allocations]
      summary: Update student allocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocationUpdate"
      responses:
        "200":
          description: Updated allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentAllocation"

  # ─── CLASSES ──────────────────────────────────────────────────────────────────
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes for institution
      parameters:
        - name: institutionId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /streams:
    get:
      operationId: listStreams
      tags: [classes]
      summary: List streams for a class
      parameters:
        - name: classId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Streams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stream"
    post:
      operationId: createStream
      tags: [classes]
      summary: Create stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamInput"
      responses:
        "201":
          description: Stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"

  # ─── DORMITORIES ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    InstitutionSummary:
      type: object
      required: [id, name, level]
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: string
        logoUrl:
          type: ["string", "null"]
...
    Application:
      type: object
      required: [id, userId, institutionId, status, step, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        institutionId:
          type: integer
        institution:
          $ref: "#/components/schemas/InstitutionSummary"
        student:
          $ref: "#/components/schemas/Student"
        status:
          type: string
          enum: [draft, pending, under_review, approved, rejected, checked_in]
        step:
          type: integer
        admissionCodeVerified:
          type: boolean
        admissionCode:
          type: ["string", "null"]
        documentsApproved:
          type: boolean
        paymentStatus:
          type: string
          enum: [unpaid, partial, paid]
        checkedInAt:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        reportingDate:
          type: ["string", "null"]
        reportingTime:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ApplicationInput:
      type: object
      required: [institutionId]
      properties:
        institutionId:
          type: integer
    ApplicationUpdate:
      type: object
      properties:
        step:
          type: integer
        status:
          type: string
        admissionCode:
          type: string
    RejectionInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    AdmissionCodeVerification:
      type: object
      required: [code, institutionId]
      properties:
        code:
          type: string
        institutionId:
          type: integer
    AdmissionVerificationResult:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        studentName:
          type: ["string", "null"]
        institutionName:
          type: ["string", "null"]
        reportingDate:
          type: ["string", "null"]
        message:
          type: ["string", "null"]

    # ─── STUDENTS ─────────────────────────────────────────────────────────────
    Student:
      type: object
      required: [id, fullName, gender, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        fullName:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        dateOfBirth:
          type: ["string", "null"]
        assessmentNumber:
          type: ["string", "null"]
        admissionNumber:
          type: ["string", "null"]
        religion:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        homeAddress:
          type: ["string", "null"]
        previousSchool:
          type: ["string", "null"]
        passportPhotoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    StudentAllocation:
      type: object
      required: [id, applicationId, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        admissionNumber:
          type: ["string", "null"]
        streamId:
          type: ["integer", "null"]
        streamName:
          type: ["string", "null"]
        dormitoryId:
          type: ["integer", "null"]
        dormitoryName:
          type: ["string", "null"]
        teacherName:
          type: ["string", "null"]
        teacherPhone:
          type: ["string", "null"]
        reportingDate:
          type: ["string", "null"]
        reportingTime:
          type: ["string", "null"]
        createdAt:
          type: string
    AllocationInput:
      type: object
      required: [applicationId]
      properties:
        applicationId:
          type: integer
        admissionNumber:
          type: string
        streamId:
          type: integer
        dormitoryId:
          type: integer
        teacherName:
          type: string
        teacherPhone:
          type: string
        reportingDate:
          type: string
        reportingTime:
          type: string
    AllocationUpdate:
      type: object
      properties:
        admissionNumber:
          type: string
        streamId:
          type: integer
        dormitoryId:
          type: integer
        teacherName:
          type: string
        teacherPhone:
          type: string
        reportingDate:
          type: string
        reportingTime:
          type: string

    # ─── CLASSES ──────────────────────────────────────────────────────────────
    Class:
      type: object
      required: [id, institutionId, name, academicYear]
      properties:
        id:
          type: integer
        institutionId:
          type: integer
        name:
          type: string
        level:
          type: ["string", "null"]
        academicYear:
          type: string
        streams:
          type: array
          items:
            $ref: "#/components/schemas/Stream"
    ClassInput:
      type: object
      required: [institutionId, name, academicYear]
      properties:
        institutionId:
          type: integer
        name:
          type: string
        level:
          type: string
        academicYear:
          type: string
    Stream:
      type: object
      required: [id, classId, name]
      properties:
        id:
          type: integer
        classId:
          type: integer
        name:
          type: string
        teacherName:
          type: ["string", "null"]
        teacherPhone:
          type: ["string", "null"]
        capacity:
          type: ["integer", "null"]
        enrolled:
          type: ["integer", "null"]
    StreamInput:
      type: object
      required: [classId, name]
      properties:
        classId:
          type: integer
        name:
          type: string
        teacherName:
          type: string
        teacherPhone:
          type: string
        capacity:
          type: integer

    # ─── DORMITORIES ──────────────────────────────────────────────────────────
...